remove lang.ts
