+++ src/common/utils/slpParser.ts
@@ -21,7 +21,7 @@ import type {
 import { ItemSpawnType } from "../types.js";
 import { Command, Frames, GameMode } from "../types.js";
 import { exists } from "./exists.js";
-import { keyBy, set } from "./lang.js";
+import { keyBy } from "./keyBy.js";
 import { RollbackCounter } from "./rollbackCounter.js";
 import { TypedEventEmitter } from "./typedEventEmitter.js";
 
@@ -378,4 +378,16 @@ export class SlpParser extends TypedEventEmitter<SlpParserEventMap> {
       }
     }
   }
+}
+
+function set<T extends Record<string, any>>(obj: T, path: (string | number)[], value: unknown): void {
+  let current: any = obj;
+  for (let i = 0; i < path.length - 1; i++) {
+    const key = path[i]!;
+    if (!(key in current)) {
+      current[key] = {};
+    }
+    current = current[key];
+  }
+  current[path[path.length - 1]!] = value;
 }
